% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MADGiC.R
\docType{data}
\name{gene_names}
\alias{gene_names}
\title{Ensembl names for all genes in the exome}
\format{one Ensembl name per line ("ENSG###########") for 19,238 lines.}
\source{
Distributed with code of Youn and Simon 2011.
}
\usage{
gene_names
}
\description{
This data set is a text file that contains all the Ensembl names
  of coding genes in build 36.
}

