\docType{data}
\name{OV.maf}
\alias{OV.maf}
\title{TCGA Ovarian MAF (Mutation Annotation Format) file}
\format{See
  \url{https://wiki.nci.nih.gov/display/TCGA/Mutation+Annotation+Format+(MAF)+Specification}
   for details about MAF files.}
\description{
  Contains all somatic mutation data from TCGA Ovarian
  project, downloaded from the TCGA data portal on October
  1, 2013 and collected into one file
}
\keyword{datasets}

