% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MADGiC.R
\docType{data}
\name{gene.rep.expr}
\alias{gene.rep.expr}
\title{Gene annotation list}
\format{Each gene is a 6 item list: 1. Ensembl ID, 2. chromosome, 3. coding
  base pairs, 4. replication timing category (1=Early, 2=Middle, 3=Late), 
  5. expression level category (1=Low, 2=Medium, 3=High), and 6. indicator
  of olfactory receptor (OR) status (1=Not an OR, 2=OR)}
\usage{
gene.rep.expr
}
\description{
This dataset contains a list with an item for each gene (18,926)
  that contains information about the basepairs, length, replication timing
  and expression level.
}

