\docType{data}
\name{gene_names}
\alias{gene_names}
\title{Ensembl names for all genes in the exome}
\format{one Ensembl name per line ("ENSG###########") for 19,238 lines.}
\source{
  Distributed with code of Youn and Simon 2011.
}
\description{
  This data set is a text file that contains all the
  Ensembl names of coding genes in build 36.
}
\keyword{datasets}

